Update the Advanced Commands Detection module
Now it's not Platforms Unifier anymore. And also it had a huge internal
upgrade, which is reflected outside too, because now all commands are
created outside the library and sent to it once only after it's loaded.
All commands are now in CmdsList, and ACD.prepareCmdsArray() is called
on ApplicationClass.
This might allow for custom and dynamic commands from now on, in case
they're needed (could be useful for contact names detection, possibly).

Also increased the buffer size on UtilsShell.executeShellCmd(). Not sure
why it wasn't on 64 already as it once was.

No need to have "stop" hardcoded anymore to stop recording audio. Actual
command implemented now (already existed on the Platforms Unifier but I
never implemented it here for some reason).