Added fragment with a list of all commands
Now it's possible to see which commands are available in the app. It's
still manual though. Would be good to automate the commands descriptions
array generation based on the CMDS_LIST array.

Fast reboot and reboot to bootloader commands added to CmdsExecutor.

Updated the ACD to version "2023-01-16 -- 01:16:49.221123 (Hora padrÃ£o
de GMT)".