Call contacts command added
Now it's possible to call the phone contacts with the "call (phone
contact) command". For that, the TelephonyManager module was created, to
keep updating the app-internal contacts list and sending the updated one
to the ACD. Also the 2 previous telephony modules are now submodules,
managed separately by this new module.

ACD version updated to "2023-01-31 -- 21:07:13.733512 (Hora padrão de
GMT)".

ModulesList class updated. Now submodules are more like modules. They
currently just are not restarted if they're stopped. But they can be
supported or not and that will be checked. They can also have their
instance stored on the array, and it can be accessed and modified by
any other class, not just the Modules Manager (for example, the new
Telephony Manager takes care of both the Phone Calls and the SMS Msgs
processors by simply calling the ModulesList functions like the Manager
or the Main Service do.