Added camera perm, fixed a few memory leaks, etc
It seems I missed the CAMERA permission. It's here only for KitKat and
below to take pictures. Above that I didn't add it to the permissions
requester/checker function because it's not used.

The Camera Manager is now properly checking if it's supported or not.

Also all modules are being properly checked if they're fully working or
not, and their destroy() methods have been completed by releasing their
resources. TTS, for example.

TelephonyManager renamed to TelephonyManagement (it's still called
Manager though) to remove the collision with the Android API with the
same name.