Hotword recognition improvement
Added more words to recognize calling VISOR. Now it's visor, visors,
wiser, and pfizer. It has detected all of these already, so seems it was
a good idea to add them.

At first I went on Google Translate and spoke "visor" to it. It detected
weiser and instantly changed it to visor. Another time it detected
pfizer. So I added those (sort of. Weiser is detected with an S on
PocketSphinx ("W AY S ER"), so I went with "wiser").

I searched for "AY Z" in the dictionary for words and added only those
that looked like "V AY Z ER" - like beginning in F or in W. And that
lead to the same Google Translate said.