More hotword recognition improvements
I've added "wiser" and "pfizer" back. And a new word of mine, "isor",
pronounced as visor but without the v (AY Z ER, in PocketSphinx
language). Changed the priorities of each on the keywords file, and
implemented a small check on onPartialResult() to check if "isor" is
contained in the 1st or 2nd hypothesis. If not (wiser and pfizer are the
only ones detected, for example. Or only wiser or only pfizer), then
ignore the detection. The reason for "isor" is that the recognizer may
miss the v sound and only hear "isor", which anyway is not a word. So
this fixes that possibility (happened already with me).