Google recognition improvements
Frozen methods checker times updated to include time for the user to
speak or for the methods to finish their actions - not just the what I
supposed would be a good time to check if the method is frozen (check
5+ seconds is too much for a method to be frozen already, and most
importantly, the user is waiting all that time to be able to call VISOR
again - not very cool).

Also fixed the speech time being so small - only 3 seconds... More than
that and the recognizer would terminate. My bad. This fix is included in
what I said above.

Also now if Google's recognizer was requested and any error occurred
while trying to start the service, VISOR will keep trying start it.

And fixed possible error when the thread would begin without
last_method_called being set - not useful.