Commands recognition improvement
Now the service is always running and all needed to do to start
listening is call startService() and it will start much faster (almost
instantaneous), because it won't have to be starting up anymore.

Sadly it can't be in a separate main process thread to stop it from
being killed by the system, but the SpeechRecognizer class can only be
used from the main process's thread for some reason.

Also improved the RECORD_MEDIA command ("start recording" added).

And corrected a bug added in the media path.