Downgrade compileSdk to SDK 31
Gomobile (for the Main Libraries) can only compile with JDK 8 on the
PATH. It also requires that the android.jar files are compiled with JDK
8. With the internal libraries, the newest android.jar files are
compiled with JDK 11 or 17. So I downgraded to SDK 31 because the
internal libraries JAR is compiled with JDK 8. It's the last one
compiled with that version - and so the last one I can use, until the
gomobile team updates gomobile or until I stop using the internal SDK
directly and without reflection (that's an option - use reflection
everywhere).