Now syncing User Settings and loading Device ones
Just like the PC app, the Android app now syncs the User Settings with
the ones from the server. In case there are none on startup, empty ones
are used and the server's credentials are requested.

Also Device Settings are now used, but not mandatory like in the PC app.