Manual APK install for Lollipop
Apparently on Lollipop it must be manual because a library is included
by Android Studio on the go, but not if the APK is generated manually
(libsupportjni.so). Lollipop doesn't have that library, at least the 5.1
version. Both on the EPSON Moverio BT-300 and on the Android Studio
emulator version. So must be generated and installed manually.