Fix settings resetting every boot
By removing directBootAware="true" from the Manifest everywhere. The
devices were trying to access the storage when it wasn't yet
available...

This means though that VISOR won't start until the user enters the PIN
(or will anyway - on BV9500 he does, who knows how). Also might mean the
PLS won't start until the user enters the PIN, which is not useful at
all, but I'll leave that to later. For now the settings are actually
working for the first time, apparently.